+++ src/modules/voxelformat/private/voxel3d/V3AFormat.cpp
@@ -1,0 +1,192 @@
+/**
+ * @file
+ */
+
+#include "V3AFormat.h"
+#include "core/Color.h"
+#include "core/Log.h"
+#include "core/StringUtil.h"
+#include "scenegraph/SceneGraph.h"
+#include "voxel/Palette.h"
+#include "voxel/PaletteLookup.h"
+
+#include <SDL_stdinc.h>
+
+namespace voxelformat {
+
+#define wrap(read)                                                                                                     \
+	if ((read) != 0) {                                                                                                 \
+		Log::error("Could not load v3a file: Not enough data in stream " CORE_STRINGIFY(read));                        \
+		return false;                                                                                                  \
+	}
+
+#define wrapBool(read)                                                                                                 \
+	if ((read) != true) {                                                                                              \
+		Log::error("Could not load v3a file: Not enough data in stream " CORE_STRINGIFY(read) " (line %i)",            \
+				   (int)__LINE__);                                                                                     \
+		return false;                                                                                                  \
+	}
+
+bool V3AFormat::loadGroupsRGBA(const core::String &filename, io::SeekableReadStream &stream,
+							   scenegraph::SceneGraph &sceneGraph, const voxel::Palette &palette,
+							   const LoadContext &ctx) {
+	core::String line;
+	int width, depth, height;
+	width = depth = height = 0;
+	while (!stream.eos()) {
+		wrapBool(stream.readLine(line));
+		if (core::string::startsWith(line, "VERSION")) {
+			line = line.substr(8);
+			if (line != "1.0") {
+				Log::error("Unsupported VERSION: %s", line.c_str());
+				return false;
+			}
+			continue;
+		}
+		if (core::string::startsWith(line, "TYPE")) {
+			line = line.substr(5);
+			if (line != "VoxelCubic") {
+				Log::error("Unsupported TYPE: %s", line.c_str());
+				return false;
+			}
+			continue;
+		}
+		if (core::string::startsWith(line, "DIMENSION")) {
+			continue;
+		}
+		if (core::string::startsWith(line, "SIZE")) {
+			line = line.substr(5);
+			if (SDL_sscanf(line.c_str(), "%i %i %i", &width, &depth, &height) != 3) {
+				Log::error("Failed to parse SIZE line: %s", line.c_str());
+				return false;
+			}
+			continue;
+		}
+		if (core::string::startsWith(line, "DATA")) {
+			line = line.substr(5);
+			break;
+		}
+		Log::warn("Unsupported line: %s", line.c_str());
+	}
+
+	if (width == 0 || height == 0 || depth == 0) {
+		Log::error("Invalid size: %i:%i:%i", width, height, depth);
+		return false;
+	}
+
+	if (width > 2048 || height > 2048 || depth > 2048) {
+		Log::error("Volume exceeds the max allowed size: %i:%i:%i", width, height, depth);
+		return false;
+	}
+
+	voxel::PaletteLookup palLookup(palette);
+	voxel::Region region(0, 0, 0, width - 1, height - 1, depth - 1);
+	voxel::RawVolume *volume = new voxel::RawVolume(region);
+	scenegraph::SceneGraphNode node;
+	node.setVolume(volume, true);
+
+	core::DynamicArray<core::String> tokens;
+	int x = 0;
+	int y = 0;
+	do {
+		if (line.empty()) {
+			if (!stream.readLine(line)) {
+				break;
+			}
+			y = 0;
+			++x;
+			if (x >= width) {
+				Log::error("Max width exceeded");
+				return false;
+			}
+		}
+
+		if (y >= height) {
+			Log::error("Max depth exceeded");
+			return false;
+		}
+		tokens.clear();
+		core::string::splitString(line, tokens);
+		if (tokens.size() % 4 != 0) {
+			Log::error("Invalid data line: %s", line.c_str());
+			return false;
+		}
+		for (size_t i = 0; i < tokens.size(); i += 4) {
+			const int r = core::string::toInt(tokens[i + 0]);
+			if (r == -1) {
+				continue;
+			}
+			const int g = core::string::toInt(tokens[i + 1]);
+			const int b = core::string::toInt(tokens[i + 2]);
+			const int a = core::string::toInt(tokens[i + 3]);
+			const core::RGBA rgba(r, g, b, a);
+			const int index = palLookup.findClosestIndex(rgba);
+			const voxel::Voxel &voxel = voxel::createVoxel(palette, index);
+			// we have to flip depth with height for our own coordinate system
+			volume->setVoxel(x, y, (int)i / 4, voxel);
+		}
+		++y;
+	} while (stream.readLine(line));
+	node.setName(filename);
+	node.setPalette(palLookup.palette());
+	sceneGraph.emplace(core::move(node));
+	return true;
+}
+
+#undef wrap
+#undef wrapBool
+
+#define wrapBool(read)                                                                                                 \
+	if ((read) != true) {                                                                                              \
+		Log::error("Could not write v3a file: " CORE_STRINGIFY(read) " (line %i)", (int)__LINE__);                     \
+		return false;                                                                                                  \
+	}
+
+bool V3AFormat::saveGroups(const scenegraph::SceneGraph &sceneGraph, const core::String &filename,
+						   io::SeekableWriteStream &stream, const SaveContext &ctx) {
+	const scenegraph::SceneGraphNode *node = sceneGraph.firstModelNode();
+	core_assert(node);
+
+	const voxel::Region &region = node->region();
+	const uint32_t width = region.getWidthInVoxels();
+	const uint32_t height = region.getHeightInVoxels();
+	const uint32_t depth = region.getDepthInVoxels();
+
+	wrapBool(stream.writeString("VERSION 1.0\r\n", false))
+	wrapBool(stream.writeString("TYPE VoxelCubic\r\n", false))
+	if (!stream.writeStringFormat(false, "DIMENSION %d.0 %d.0 %d.0\r\n", width, height, depth)) {
+		Log::error("Failed to write DIMENSION line");
+		return false;
+	}
+	if (!stream.writeStringFormat(false, "SIZE %d %d %d\r\n", width, height, depth)) {
+		Log::error("Failed to write SIZE line");
+		return false;
+	}
+	wrapBool(stream.writeString("DATA ", false))
+	voxel::RawVolume &volume = *node->volume();
+	const voxel::Palette &palette = node->palette();
+	for (int32_t x = region.getLowerX(); x <= region.getUpperX(); x++) {
+		for (int32_t y = region.getLowerY(); y <= region.getUpperY(); y++) {
+			for (int32_t z = region.getLowerZ(); z <= region.getUpperZ(); z++) {
+				const voxel::Voxel &voxel = volume.voxel(x, y, z);
+				if (voxel::isAir(voxel.getMaterial())) {
+					wrapBool(stream.writeString("-1 -1 -1 -1 ", false))
+				} else {
+					const core::RGBA color = palette.color(voxel.getColor());
+					if (!stream.writeStringFormat(false, "%d %d %d %d ", color.r, color.g, color.b, color.a)) {
+						Log::error("Failed to write voxel data");
+						return false;
+					}
+				}
+			}
+			wrapBool(stream.writeString("\r\n", false))
+		}
+		wrapBool(stream.writeString("\r\n", false))
+	}
+
+	return true;
+}
+
+#undef wrapBool
+
+} // namespace voxelformat