+++ src/modules/voxelformat/private/voxel3d/V3AFormat.h
@@ -1,0 +1,30 @@
+/**
+ * @file
+ */
+
+#pragma once
+
+#include "voxelformat/Format.h"
+
+namespace voxelformat {
+
+/**
+ * @brief EveryGraph Voxel3D format
+ *
+ * @ingroup Formats
+ */
+class V3AFormat : public RGBAFormat {
+protected:
+	bool loadGroupsRGBA(const core::String &filename, io::SeekableReadStream &stream,
+						scenegraph::SceneGraph &sceneGraph, const voxel::Palette &palette,
+						const LoadContext &ctx) override;
+	bool saveGroups(const scenegraph::SceneGraph &sceneGraph, const core::String &filename,
+					io::SeekableWriteStream &stream, const SaveContext &ctx) override;
+
+public:
+	bool singleVolume() const override {
+		return true;
+	}
+};
+
+} // namespace voxelformat