+++ docs/CHANGELOG.md
@@ -30,6 +30,7 @@ General:
    - Expose shape generators to lua
    - Fixed color intensity handling for `kvx` files
    - Added write support for `kvx` format (used in e.g. voxel doom and eduke3d)
+   - Added support for Voxel3D `v3a` format
 
 VoxEdit: